js: split into js.py
